// src/types/sportdb.ts
export interface Standing {
    againstPctg: string;
    draws: string;
    eventTournamentStageCaptionPart2: string;
    goalDiff: string;
    goals: string;

    lossesAfterPenalties: string;
    lossesInOvertimeAfterPenalties: string;
    lossesOvertime: string;
    lossesRegular: string;

    matches: string;
    noResultsMatches: string;

    points: string;
    pointsPerMatchesPlayed: string;

    rank: string;
    rankClass: string;
    rankColor: string;

    teamId: string;
    teamName: string;
    teamSlug: string;
    teamUrl: string;

    wins: string;
    winsAfterPenalties: string;
    winsInOvertimeAfterPenalties: string;
    winsOvertime: string;
    winsRegular: string;

    events: unknown[];
}